+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add tests for useFetchExportTitlesFromPackage hook. (UIEH-1182)
 * Replace or remove react-hot-loader. (UIEH-1265)
 * Package Detail Record> Usage & analysis accordion > Apply pagination to Titles list. (UIEH-1259)
+* Remove call to "/groups" api in settings/eholdings/assigned users. (UIEH-1236)
 
 ## [7.1.4] (https://github.com/folio-org/ui-eholdings/tree/v7.1.4) (2022-04-08)